Split out a `tracing_web.rs` module
Use different source files for the `tracing` module depending on the
target.
