+++ linera-base/src/lib.rs
@@ -24,6 +24,7 @@ pub mod prometheus_util;
 #[cfg(not(chain))]
 pub mod task;
 pub mod time;
+#[cfg_attr(web, path = "tracing_web.rs")]
 pub mod tracing;
 #[cfg(test)]
 mod unit_tests;
